refactor : 폴더 구조 변경 - config 패키지에서 properties 패키지 분리 - UserDetails 관련 파일 jwt 패키지에서 domain.member 패키지로 분리
