+++ src/main/java/com/project/singk/domain/member/service/AuthService.java
@@ -11,12 +11,12 @@ import com.project.singk.domain.member.dto.SignupRequestDto;
 import com.project.singk.domain.member.repository.MemberRepository;
 import com.project.singk.global.api.ApiException;
 import com.project.singk.global.api.AppHttpStatus;
-import com.project.singk.global.config.properties.JwtProperties;
-import com.project.singk.global.config.properties.MailProperties;
+import com.project.singk.global.properties.JwtProperties;
+import com.project.singk.global.properties.MailProperties;
 import com.project.singk.global.domain.PkResponseDto;
 import com.project.singk.global.domain.TokenDto;
 import com.project.singk.global.jwt.JwtUtil;
-import com.project.singk.global.jwt.SingKUserDetails;
+import com.project.singk.domain.member.domain.SingKUserDetails;
 import com.project.singk.global.util.RedisUtil;
 
 import io.jsonwebtoken.Claims;
