+++ src/main/java/com/project/singk/domain/member/service/SingKUserDetailsService.java
@@ -1,10 +1,11 @@
-package com.project.singk.global.jwt;
+package com.project.singk.domain.member.service;
 
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
+import com.project.singk.domain.member.domain.SingKUserDetails;
 import com.project.singk.domain.member.repository.MemberRepository;
 import com.project.singk.global.api.ApiException;
 import com.project.singk.global.api.AppHttpStatus;
