+++ src/main/java/com/project/singk/global/properties/CorsProperties.java
@@ -1,4 +1,4 @@
-package com.project.singk.global.config.properties;
+package com.project.singk.global.properties;
 
 import java.util.List;
